[IMP] Add comments, generalize create_csv_report
