+++ sale_order_export_rpl/report/__init__.py
@@ -1,2 +1,3 @@
+from . import report_custom_csv
 from . import report_sale_order_csv
 from . import report_sale_order_line_csv
