+++ sale_order_export_rpl/report/report_sale_order_line_csv.py
@@ -2,21 +2,12 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import csv
-from io import StringIO
-
 from odoo import models
 
 
 class SaleOrderLineCSV(models.AbstractModel):
     _name = "report.report_csv.sale_order_line_csv"
-    _inherit = "report.report_csv.abstract"
-
-    def create_csv_report(self, docids, data):
-        objs = self._get_objs_for_report(docids, data)
-        file_data = StringIO()
-        self.generate_csv_report(data, objs, file_data)
-        file_data.seek(0)
-        return file_data.read(), "csv"
+    _inherit = "report.report_csv.custom"
 
     def generate_csv_report(self, data, orders, file_data):
         writer = csv.DictWriter(
